pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doris-mcp"
version = "0.2.0"
description = "Doris MCP Server for Cursor integration"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
authors = [
    {name = "Doris MCP Team - Yijia Su"}
]

dependencies = [
    "mcp[cli]>=1.0.0",
    "pymysql>=1.0.2",
    "pandas>=1.5.0",
    "numpy>=1.20.0",
    "scikit-learn>=1.0.0",
    "python-dotenv>=0.19.0",
    "pydantic>=1.10.0",
    "requests>=2.28.0", 
    "openai>=1.66.3",
    "fastapi>=0.95.0",
    "uvicorn>=0.21.0",
    "simplejson>=3.17.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0"
]

[project.scripts]
# Updated entry point for stdio mode back to mcp_core
doris-mcp = "doris_mcp_server.mcp_core:run_stdio"

[tool.setuptools]
# Explicitly list the package found in the root directory
packages = ["doris_mcp_server"]
